third_party/blink/web_tests/platform/generic/http/tests/websocket/zero-length-text-expected.txt

Test whether WebSocket handles zero length text frame correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onmessage() was called. (message = "")
onmessage() was called. (message = "This first text should be received.")
onmessage() was called. (message = "")
onmessage() was called. (message = "This second text should be received, too.")
onclose() was called.
PASS closeEvent.wasClean is true
PASS sequence is 4
PASS receivedText is expectedText
PASS successfullyParsed is true

TEST COMPLETE
